pyproject.toml
[project]
name = "async-rpc"
version = "0.1.0"
description = "Resonate Async RPC example application"
authors = [
    { name = "Resonate HQ", email = "contact@resoantehq.io" }
]
dependencies = [
    "flask>=3.1.0",
    "resonate-sdk>=0.5.0"
]

readme = "README.md"
requires-python = ">=3.12"

[project.scripts]
gateway = "src.gateway:main"
a = "src.service_a:main"
b = "src.service_b:main"
c = "src.service_c:main"
d = "src.service_d:main"
e = "src.service_e:main"
f = "src.service_f:main"
g = "src.service_g:main"
h = "src.service_h:main"
i = "src.service_i:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]